requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.8.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
cytoolz==0.12.2
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
django-rest-knox==4.2.0
django-sendgrid-v5==1.2.2
djangorestframework==3.14.0
drf-extra-fields==3.5.0
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
hexbytes==0.3.1
idna==3.4
importlib-metadata==6.8.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keras==2.13.1
libclang==16.0.6
lru-dict==1.2.0
Markdown==3.4.4
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
parsimonious==0.9.0
Pillow==10.0.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pytesseract==0.3.10
python-http-client==3.3.7
pytz==2023.3
pyunormalize==15.0.0
pywin32==306
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rlp==3.0.0
rpds-py==0.9.2
rsa==4.9
sendgrid==6.10.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
toolz==0.12.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
web3==6.7.0
websockets==11.0.3
Werkzeug==2.3.6
word-beam-search @ file:///F:/eth_id_db/venv/Lib/site-packages/CTCWordBeamSearch
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
